+++ Vineyard/Interface/IParcelService.cs
@@ -5,6 +5,6 @@
         ICollection<Parcel> GetAll();
         ICollection<Parcel> GetByGrape(string grape);
         Parcel GetById(int id);
-        long RecommendWateringAmount(int parcelId);
+        double RecommendWateringAmount(int parcelId);
     }
 }
